Update folder name in serverless-example-s3.md
Update "Upload an image to the Amazon S3 bucket that you created for this sample application." to "Upload an image to the Amazon S3 bucket named `aws-sam-ocr-sourceimagebucket` that was created by this sample application.". The S3 bucket that the image needs to be uploaded to is not the S3 bucket you created. It is the S3 bucket referenced in the template.yaml of the example AWS SAM project.